+++ src/main/drivers/telemetry.c
@@ -47,7 +47,7 @@ void telemetry(void) {
   } else {
     telemetryData.current = 0;
   }
-  telemetryData.consumption = consumptionMah;
+  telemetryData.consumption = (int)consumptionMah;
   telemetryData.erpm = 542137.4/motor.CommutationInterval;
 
   telemetryTelegram(&telemetryData);
